Cleanup inventory wrapper implementation, add < 64 max stack count test, separate tests better
